requirements/service-requirements.txt
Django==2.1
djangorestframework==3.8.2
requests==2.19.1
WeasyPrint==0.42.3
bleach==2.1.3